+++ src/dune_rules/compilation_context.ml
@@ -3,6 +3,8 @@ open! Stdune
 open Import
 module SC = Super_context
 
+let modules_of_lib = Fdecl.create Dyn.Encoder.opaque
+
 module Includes = struct
   type t = Command.Args.dynamic Command.Args.t Cm_kind.Dict.t
 
@@ -164,8 +166,8 @@ let for_alias_module t =
   let flags =
     let project = Scope.project t.scope in
     let dune_version = Dune_project.dune_version project in
-    Ocaml_flags.default ~profile:(Super_context.context t.super_context).profile
-      ~dune_version
+    let profile = (Super_context.context t.super_context).profile in
+    Ocaml_flags.default ~dune_version ~profile
   in
   let sandbox =
     let ctx = Super_context.context t.super_context in
@@ -184,6 +186,20 @@ let for_alias_module t =
   ; includes = Includes.empty
   ; stdlib = None
   ; sandbox
+  }
+
+let for_root_module t =
+  let flags =
+    let project = Scope.project t.scope in
+    let dune_version = Dune_project.dune_version project in
+    let profile = (Super_context.context t.super_context).profile in
+    Ocaml_flags.default ~profile ~dune_version
+  in
+  { t with
+    flags =
+      Ocaml_flags.append_common flags
+        [ "-w"; "-49"; "-nopervasives"; "-nostdlib" ]
+  ; stdlib = None
   }
 
 let for_module_generated_at_link_time cctx ~requires ~module_ =
@@ -214,3 +230,9 @@ let for_plugin_executable t ~embed_in_plugin_libraries =
   { t with requires_link }
 
 let without_bin_annot t = { t with bin_annot = false }
+
+let root_module_entries t : Module_name.t list Or_exn.t =
+  let open Result.O in
+  let* requires = t.requires_compile in
+  let local_lib = Fdecl.get modules_of_lib t.super_context in
+  Result.List.concat_map requires ~f:(Lib.entry_module_names ~local_lib)