+++ src/dune_rules/dune_file.ml
@@ -45,6 +45,10 @@ module Js_of_ocaml = struct
     { flags = Ordered_set_lang.Unexpanded.standard; javascript_files = [] }
 end
 
+type for_ =
+  | Executable
+  | Library of Wrapped.t option
+
 module Lib_deps = struct
   type t = Lib_dep.t list
 
@@ -53,9 +57,16 @@ module Lib_deps = struct
     | Optional
     | Forbidden
 
-  let decode ~allow_re_export =
+  let decode for_ =
     let+ loc = loc
-    and+ t = repeat (Lib_dep.decode ~allow_re_export) in
+    and+ t =
+      let allow_re_export =
+        match for_ with
+        | Library _ -> true
+        | Executable -> false
+      in
+      repeat (Lib_dep.decode ~allow_re_export)
+    in
     let add kind name acc =
       match Lib_name.Map.find acc name with
       | None -> Lib_name.Map.set acc name kind
@@ -103,7 +114,7 @@ module Lib_deps = struct
   let info t ~kind =
     List.concat_map t ~f:(function
       | Lib_dep.Re_export (_, s)
-      | Lib_dep.Direct (_, s) ->
+      | Direct (_, s) ->
         [ (s, kind) ]
       | Select { choices; _ } ->
         List.concat_map choices ~f:(fun (c : Lib_dep.Select.Choice.t) ->
@@ -161,18 +172,18 @@ module Buildable = struct
     ; flags : Ocaml_flags.Spec.t
     ; js_of_ocaml : Js_of_ocaml.t
     ; allow_overlapping_dependencies : bool
+    ; root_module : (Loc.t * Module_name.t) option
     }
 
-  let decode ~in_library ~allow_re_export =
+  let decode (for_ : for_) =
     let use_foreign =
       Dune_lang.Syntax.deleted_in Stanza.syntax (2, 0)
         ~extra_info:"Use the (foreign_stubs ...) field instead."
     in
     let only_in_library decode =
-      if in_library then
-        decode
-      else
-        return None
+      match for_ with
+      | Executable -> return None
+      | Library _ -> decode
     in
     let add_stubs language ~loc ~names ~flags foreign_stubs =
       match names with
@@ -219,8 +230,7 @@ module Buildable = struct
               >>> enter (maybe string) )))
     and+ modules_without_implementation =
       Stanza_common.modules_field "modules_without_implementation"
-    and+ libraries =
-      field "libraries" (Lib_deps.decode ~allow_re_export) ~default:[]
+    and+ libraries = field "libraries" (Lib_deps.decode for_) ~default:[]
     and+ flags = Ocaml_flags.Spec.decode
     and+ js_of_ocaml =
       field "js_of_ocaml" Js_of_ocaml.decode ~default:Js_of_ocaml.default
@@ -254,6 +264,9 @@ module Buildable = struct
                        repeat (String_with_vars.decode >>| version_check)
                      in
                      (libname, flags))) ))
+    and+ root_module =
+      field_o "root_module"
+        (Dune_lang.Syntax.since Stanza.syntax (2, 8) >>> Module_name.decode_loc)
     in
     let preprocess =
       let init =
@@ -309,6 +322,7 @@ module Buildable = struct
     ; flags
     ; js_of_ocaml
     ; allow_overlapping_dependencies
+    ; root_module
     }
 
   let has_foreign t =
@@ -564,8 +578,9 @@ module Library = struct
   let decode =
     fields
       (let* stanza_loc = loc in
+       let* wrapped = Wrapped.field in
        let* dune_version = Dune_lang.Syntax.get_exn Stanza.syntax in
-       let+ buildable = Buildable.decode ~in_library:true ~allow_re_export:true
+       let+ buildable = Buildable.decode (Library (Option.map ~f:snd wrapped))
        and+ name = field_o "name" Lib_name.Local.decode_loc
        and+ public =
          field_o "public_name" (Public_lib.decode ~allow_deprecated_names:false)
@@ -585,7 +600,6 @@ module Library = struct
          field "modes" Mode_conf.Set.decode
            ~default:(Mode_conf.Set.default stanza_loc)
        and+ kind = field "kind" Lib_kind.decode ~default:Lib_kind.Normal
-       and+ wrapped = Wrapped.field
        and+ optional = field_b "optional"
        and+ no_dynlink = field_b "no_dynlink"
        and+ () =
@@ -902,13 +916,14 @@ module Library = struct
     let wrapped = Some conf.wrapped in
     let special_builtin_support = conf.special_builtin_support in
     let instrumentation_backend = conf.instrumentation_backend in
+    let entry_modules = Lib_info.Source.Local in
     Lib_info.create ~loc ~name ~kind ~status ~src_dir ~orig_src_dir ~obj_dir
       ~version ~synopsis ~main_module_name ~sub_systems ~requires
       ~foreign_objects ~plugins ~archives ~ppx_runtime_deps ~foreign_archives
       ~native_archives ~foreign_dll_files ~jsoo_runtime ~jsoo_archive
-      ~preprocess ~enabled ~virtual_deps ~dune_version ~virtual_ ~implements
-      ~default_implementation ~modes ~wrapped ~special_builtin_support
-      ~exit_module ~instrumentation_backend
+      ~preprocess ~enabled ~virtual_deps ~dune_version ~virtual_ ~entry_modules
+      ~implements ~default_implementation ~modes ~wrapped
+      ~special_builtin_support ~exit_module ~instrumentation_backend
 end
 
 module Plugin = struct
@@ -1347,7 +1362,7 @@ module Executables = struct
 
   let common =
     let* dune_version = Dune_lang.Syntax.get_exn Stanza.syntax in
-    let+ buildable = Buildable.decode ~in_library:false ~allow_re_export:false
+    let+ buildable = Buildable.decode Executable
     and+ (_ : bool) =
       field "link_executables" ~default:true
         (Dune_lang.Syntax.deleted_in Stanza.syntax (1, 0) >>> bool)
@@ -1772,8 +1787,7 @@ module Tests = struct
 
   let gen_parse names =
     fields
-      (let+ buildable =
-         Buildable.decode ~in_library:false ~allow_re_export:false
+      (let+ buildable = Buildable.decode Executable
        and+ link_flags = Ordered_set_lang.Unexpanded.field "link_flags"
        and+ names = names
        and+ package = field_o "package" Stanza_common.Pkg.decode