+++ src/dune_rules/lib_dep.mli
@@ -30,11 +30,7 @@ val direct : Loc.t * Lib_name.t -> t
 
 val re_export : Loc.t * Lib_name.t -> t
 
-val to_lib_names : t -> Lib_name.t list
-
 val decode : allow_re_export:bool -> t Dune_lang.Decoder.t
-
-val encode : t Dune_lang.Encoder.t
 
 module L : sig
   val field_encode : t list -> name:string -> Dune_lang.Encoder.field